+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 
 ## DEPENDENCIES
 
+- **eslint** has been upgraded from 22.0.0 to **23.0.0**
 - **grunt-contrib-compass** has been removed
 - **grunt-contrib-concat** has been removed
 - **grunt-contrib-uglify** has been removed
 - **grunt-contrib-watch** has been removed
+- **grunt-eslint** has been removed
 - **grunt-mocha-test** has been removed
 - **grunt-spritesmith** has been removed
 - **interactjs** has been upgraded from 1.9.7 to **1.10.11**